String to int conversion without atoi
Code to swich string to int. Logic is commented. It's here if you guys think we should use it. I don't know if you guys want to stick withall of the code seperate or go back to one file where everything worked. Either way the code can just be copied and pasted.